Moving custom pivot updating events to existing test class
