+++ tests/Integration/Database/EloquentPivotEventsTest.php
@@ -31,6 +31,7 @@ class EloquentPivotEventsTest extends DatabaseTestCase
         Schema::create('project_users', function (Blueprint $table) {
             $table->integer('user_id');
             $table->integer('project_id');
+            $table->text('permissions')->nullable();
             $table->string('role')->nullable();
         });
 
@@ -73,6 +74,52 @@ class EloquentPivotEventsTest extends DatabaseTestCase
         $project->contributors()->detach($user->id);
         $this->assertEquals([], PivotEventsTestCollaborator::$eventsCalled);
     }
+
+    public function testCustomPivotUpdateEventHasExistingAttributes()
+    {
+        $_SERVER['pivot_attributes'] = false;
+
+        $user = PivotEventsTestUser::forceCreate([
+            'email' => 'taylor@laravel.com',
+        ]);
+
+        $project = PivotEventsTestProject::forceCreate([
+            'name' => 'Test Project',
+        ]);
+
+        $project->collaborators()->attach($user, ['permissions' => ['foo', 'bar']]);
+
+        $project->collaborators()->updateExistingPivot($user->id, ['role' => 'Lead Developer']);
+
+        $this->assertSame(
+            [
+                'user_id' => '1',
+                'project_id' => '1',
+                'permissions' => '["foo","bar"]',
+                'role' => 'Lead Developer',
+            ],
+            $_SERVER['pivot_attributes']
+        );
+    }
+
+    public function testCustomPivotUpdateEventHasDirtyCorrect()
+    {
+        $_SERVER['pivot_dirty_attributes'] = false;
+
+        $user = PivotEventsTestUser::forceCreate([
+            'email' => 'taylor@laravel.com',
+        ]);
+
+        $project = PivotEventsTestProject::forceCreate([
+            'name' => 'Test Project',
+        ]);
+
+        $project->collaborators()->attach($user, ['permissions' => ['foo', 'bar'], 'role' => 'Developer']);
+
+        $project->collaborators()->updateExistingPivot($user->id, ['role' => 'Lead Developer']);
+
+        $this->assertSame(['role' => 'Lead Developer'], $_SERVER['pivot_dirty_attributes']);
+    }
 }
 
 class PivotEventsTestUser extends Model
@@ -103,6 +150,10 @@ class PivotEventsTestCollaborator extends Pivot
 {
     public $table = 'project_users';
 
+    protected $casts = [
+        'permissions' => 'json',
+    ];
+
     public static $eventsCalled = [];
 
     public static function boot()
@@ -122,6 +173,8 @@ class PivotEventsTestCollaborator extends Pivot
         });
 
         static::updated(function ($model) {
+            $_SERVER['pivot_attributes'] = $model->getAttributes();
+            $_SERVER['pivot_dirty_attributes'] = $model->getDirty();
             static::$eventsCalled[] = 'updated';
         });
